Chore: Move VideoImageHasher into hashing feature folder
