+++ lib/features/hashing/logic/ImageManger.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 import 'dart:typed_data';
-import 'package:decentproof/pages/videoimagepage/logic/ImagePickerWrapper.dart';
+import 'package:decentproof/features/hashing/logic/ImagePickerWrapper.dart';
 import 'package:decentproof/shared/ExifWrapper.dart';
 import 'package:nanoid/async.dart';
 import 'package:path_provider/path_provider.dart';
