+++ lib/features/hashing/uiblocks/ImageButton.dart
@@ -1,17 +1,17 @@
-import 'package:decentproof/pages/videoimagepage/logic/VideoImageHashManager.dart';
-import 'package:decentproof/shared/ProcessingDialog.dart';
+import 'package:decentproof/features/hashing/logic/hasher/VideoImageHashManager.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import '../logic/VideoManager.dart';
 
-class VideoButton extends StatelessWidget {
-  const VideoButton(
+import '../../../shared/ProcessingDialog.dart';
+import '../logic/ImageManger.dart';
+
+class ImageButton extends StatelessWidget {
+  const ImageButton(
       {Key? key,
-      required this.videoManager,
+      required this.imageManager,
       required this.videoImageHashManager})
       : super(key: key);
-  final VideoManager videoManager;
+  final ImageManager imageManager;
   final VideoImageHashManager videoImageHashManager;
   @override
   Widget build(BuildContext context) {
@@ -26,18 +26,13 @@ class VideoButton extends StatelessWidget {
           color: Colors.transparent,
           child: InkWell(
             onTap: () async {
+              String path = await imageManager.saveImage();
               showDialog(
                   context: context,
                   builder: (context) => const ProcessingDialog());
-              try {
-                String path = await videoManager.saveVideo();
-                String hash =
-                    await compute(videoImageHashManager.hashVideo, path);
-                Navigator.of(context).pushNamed("/submissionPage",
-                    arguments: {"hash": hash, "path": path});
-              } catch (e) {
-                Navigator.of(context).pop();
-              }
+              String hash = await videoImageHashManager.hashImage(path);
+              Navigator.of(context).pushNamed("/submissionPage",
+                  arguments: {"hash": hash, "path": path});
             },
             radius: size.width,
             splashColor: Colors.orangeAccent,
@@ -45,13 +40,12 @@ class VideoButton extends StatelessWidget {
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 children: [
                   const Icon(
-                    Icons.video_file,
+                    Icons.camera,
                     size: 100.0,
                   ),
-                  const Text("videoImagePage.video",
-                          style: TextStyle(
-                              fontWeight: FontWeight.bold, fontSize: 20.0))
-                      .tr()
+                  Text("videoImagePage.photo".tr(),
+                      style: const TextStyle(
+                          fontWeight: FontWeight.bold, fontSize: 20.0))
                 ]),
           )),
     );
