+++ lib/features/hashing/uiblocks/VideoButton.dart
@@ -1,17 +1,17 @@
-import 'package:decentproof/pages/videoimagepage/logic/ImageManger.dart';
-import 'package:decentproof/pages/videoimagepage/logic/VideoImageHashManager.dart';
+import 'package:decentproof/features/hashing/logic/hasher/VideoImageHashManager.dart';
+import 'package:decentproof/shared/ProcessingDialog.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import '../logic/VideoManager.dart';
 
-import '../../../shared/ProcessingDialog.dart';
-
-class ImageButton extends StatelessWidget {
-  const ImageButton(
+class VideoButton extends StatelessWidget {
+  const VideoButton(
       {Key? key,
-      required this.imageManager,
+      required this.videoManager,
       required this.videoImageHashManager})
       : super(key: key);
-  final ImageManager imageManager;
+  final VideoManager videoManager;
   final VideoImageHashManager videoImageHashManager;
   @override
   Widget build(BuildContext context) {
@@ -26,13 +26,18 @@ class ImageButton extends StatelessWidget {
           color: Colors.transparent,
           child: InkWell(
             onTap: () async {
-              String path = await imageManager.saveImage();
               showDialog(
                   context: context,
                   builder: (context) => const ProcessingDialog());
-              String hash = await videoImageHashManager.hashImage(path);
-              Navigator.of(context).pushNamed("/submissionPage",
-                  arguments: {"hash": hash, "path": path});
+              try {
+                String path = await videoManager.saveVideo();
+                String hash =
+                    await compute(videoImageHashManager.hashVideo, path);
+                Navigator.of(context).pushNamed("/submissionPage",
+                    arguments: {"hash": hash, "path": path});
+              } catch (e) {
+                Navigator.of(context).pop();
+              }
             },
             radius: size.width,
             splashColor: Colors.orangeAccent,
@@ -40,12 +45,13 @@ class ImageButton extends StatelessWidget {
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 children: [
                   const Icon(
-                    Icons.camera,
+                    Icons.video_file,
                     size: 100.0,
                   ),
-                  Text("videoImagePage.photo".tr(),
-                      style: const TextStyle(
-                          fontWeight: FontWeight.bold, fontSize: 20.0))
+                  const Text("videoImagePage.video",
+                          style: TextStyle(
+                              fontWeight: FontWeight.bold, fontSize: 20.0))
+                      .tr()
                 ]),
           )),
     );
